Evil circle & Ball count
At the point where you loop through every ball and call the draw(), update(), and collisionDetect() functions for each one, make it so that these functions are only called if the current ball exists.
Call the evil circle instance's draw(), checkBounds(), and collisionDetect() methods on every iteration of the loop.